+++ geowave/benchmark/src/main/scala/geotrellis/geowave/QueryBenchmark.scala
@@ -1,3 +1,19 @@
+/*
+ * Copyright 2021 Azavea
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
 package geotrellis.geowave
 
 import geotrellis.geowave.dsl.syntax._
@@ -7,6 +23,9 @@ import geotrellis.geowave.index.query.ExplicitSpatialTemporalElevationQuery
 import geotrellis.geowave.adapter.TypeName
 import geotrellis.geowave.api._
 import geotrellis.vector._
+import geotrellis.raster.io.geotiff.GeoTiff
+import geotrellis.raster.MultibandTile
+import geotrellis.store.util.BlockingThreadPool
 
 import org.locationtech.geowave.core.geotime.index.api.{SpatialIndexBuilder, SpatialTemporalIndexBuilder}
 import org.locationtech.geowave.core.store.api.{DataStore, DataStoreFactory, Index}
@@ -24,8 +43,6 @@ import scala.collection.JavaConverters._
 import java.util.concurrent.TimeUnit
 import java.time.{LocalDate, ZoneOffset}
 import java.util.Date
-import geotrellis.raster.io.geotiff.GeoTiff
-import geotrellis.raster.MultibandTile
 
 @BenchmarkMode(Array(Mode.AverageTime))
 @State(Scope.Benchmark)